+++ docs/changelog.md
@@ -15,6 +15,7 @@ Release notes for `quimb`.
 **Bug fixes:**
 
 - Fix [`TensorNetwork2D.compute_plaquette_environments`](quimb.tensor.tensor_2d.TensorNetwork2D.compute_plaquette_environments) for `mode="zipup"` and other boundary contraction methods that use the generic 1D compression algorithms.
+- [`parse_openqasm2_str`](quimb.tensor.circuit.parse_openqasm2_str) allow custom gate names to start with the word `gate` ({issue}`312`).
 
 
 (whats-new-1-11-0)=